backend/requirements.txt
Flask==3.1.0
flask-cors==6.0.1
gunicorn==23.0.0
openai==1.85.0
python-dotenv==1.1.1
pypdf==5.1.0
werkzeug==3.1.3